fix: corrigir erros dbt e healthcheck no docker-compose
- dbt/dbt_project.yml: adiciona collected_at e inserted_at como
  timestamptz no column_types do seed historical_prices para resolver
  UNION type mismatch (timestamptz vs text) em 5 modelos mart

- agg_condition_distribution, agg_market_overview, agg_price_evolution:
  adiciona ::numeric após percentile_cont() para resolver erro
  "function round(double precision, integer) does not exist" no PostgreSQL

- stg_products.yml: corrige sintaxe depreciada do accepted_values
  adicionando nível arguments: conforme dbt-utils moderno

- docker-compose.yml: substitui healthcheck do prefect-server de curl
  (não instalado na imagem prefecthq/prefect:2-latest) por python urllib

+++ dbt/models/marts/agg_condition_distribution.sql
@@ -61,7 +61,7 @@ by_condition as (
         min(price)                                              as min_price,
         max(price)                                              as max_price,
         round(percentile_cont(0.50)
-              within group (order by price), 2)                 as median_price,
+              within group (order by price)::numeric, 2)        as median_price,
 
         -- Desconto
         round(avg(discount_pct),     2)                         as avg_discount_pct,
